+++ internal/pkg/cmd/cli/kapps/vars.go
@@ -22,6 +22,7 @@ type varsConfig struct {
 	account         string
 	cluster         string
 	region          string
+	includeParents  bool
 	includeSelector []string
 	excludeSelector []string
 	suppress        []string
@@ -35,22 +36,23 @@ func newVarsCmd(out io.Writer) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "vars [flags] [stack-file] [stack-name]",
 		Short: fmt.Sprintf("Display all variables available for a kapp"),
-		Long: `Merges variables from all sources and displays them. If the path 
-to a cache directory is given (with the '--cache-dir' flag) each kapp's sugarkube.yaml 
-file will be displayed templated.`,
+		Long: `Merges variables from all sources and displays them along with each kapp's 
+templated sugarkube.yaml file.`,
 		RunE: func(cmd *cobra.Command, args []string) error {
-			if len(args) < 2 {
-				return errors.New("the name of the stack to run, and the path to the stack file are required")
-			} else if len(args) > 2 {
+			if len(args) < 3 {
+				return errors.New("some required arguments are missing")
+			} else if len(args) > 3 {
 				return errors.New("too many arguments supplied")
 			}
 			c.stackFile = args[0]
 			c.stackName = args[1]
+			c.cacheDir = args[2]
 			return c.run()
 		},
 	}
 
 	f := cmd.Flags()
+	f.BoolVar(&c.includeParents, "parents", false, "process all parents of all selected kapps as well")
 	f.StringVar(&c.provider, "provider", "", "name of provider, e.g. aws, local, etc.")
 	f.StringVar(&c.provisioner, "provisioner", "", "name of provisioner, e.g. kops, minikube, etc.")
 	f.StringVar(&c.profile, "profile", "", "launch profile, e.g. dev, test, prod, etc.")
@@ -87,66 +89,16 @@ func (c *varsConfig) run() error {
 	}
 
 	dagObj, err := BuildDagForSelected(stackObj, c.cacheDir, c.includeSelector, c.excludeSelector,
-		"", c.out)
+		c.includeParents, "", c.out)
 	if err != nil {
 		return errors.WithStack(err)
 	}
 
-	// todo - delete this and update the dag executor to display variables
 	dagObj = dagObj
-
-	//for _, kappObj := range selectedInstallables {
-	//	templatedVars, err := stackObj.GetTemplatedVars(kappObj, map[string]interface{}{})
-	//	if err != nil {
-	//		return errors.WithStack(err)
-	//	}
-	//
-	//	if len(c.suppress) > 0 {
-	//		for _, exclusion := range c.suppress {
-	//			// trim any leading zeroes for compatibility with how variables are referred to in templates
-	//			exclusion = strings.TrimPrefix(exclusion, ".")
-	//			blanked := datautils.BlankNestedMap(map[string]interface{}{}, strings.Split(exclusion, "."))
-	//			log.Logger.Debugf("blanked=%#v", blanked)
-	//
-	//			err = mergo.Merge(&templatedVars, blanked, mergo.WithOverride)
-	//			if err != nil {
-	//				return errors.WithStack(err)
-	//			}
-	//		}
-	//	}
-	//
-	//	yamlData, err := yaml.Marshal(&templatedVars)
-	//	if err != nil {
-	//		return errors.WithStack(err)
-	//	}
-	//
-	//	_, err = fmt.Fprintf(c.out, "\n***** Start variables for kapp '%s' *****\n"+
-	//		"%s***** End variables for kapp '%s' *****\n",
-	//		kappObj, yamlData, kappObj)
-	//	if err != nil {
-	//		return errors.WithStack(err)
-	//	}
-	//
-	//	if c.cacheDir == "" {
-	//		_, err = fmt.Fprintf(c.out, "Won't display the %s file for "+
-	//			"'%s'. Provide the path to the cache dir with the --cache-dir "+
-	//			"option to display it.\n", kappObj, constants.KappConfigFileName)
-	//	} else {
-	//		err = kappObj.TemplateDescriptor(templatedVars)
-	//		if err != nil {
-	//			return errors.WithStack(err)
-	//		}
-	//
-	//		kappConfig, err := yaml.Marshal(kappObj.GetDescriptor())
-	//		if err != nil {
-	//			return errors.WithStack(err)
-	//		}
-	//
-	//		_, err = fmt.Fprintf(c.out, "\n***** Start config for kapp '%s' *****\n"+
-	//			"%s***** End config for kapp '%s' *****\n",
-	//			kappObj.FullyQualifiedId(), kappConfig, kappObj.FullyQualifiedId())
-	//	}
-	//}
+	err = dagObj.ExecuteGetVars(constants.DagActionVars, stackObj, c.suppress)
+	if err != nil {
+		return errors.WithStack(err)
+	}
 
 	return nil
 }